requirements.txt
kaggle
pandas
numpy
boto3
python-dotenv
pyarrow
fastparquet
google-cloud-bigquery
